requirements.txt
# Core Deep Learning
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0

# Data Processing
numpy>=1.24.0
pandas>=2.0.0
pillow>=9.5.0
opencv-python>=4.7.0
albumentations>=1.3.0
scikit-learn>=1.3.0

# Metrics & Evaluation
nltk>=3.8.0
rouge-score>=0.1.2
sacrebleu>=2.3.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.14.0
tensorboard>=2.13.0

# Explainability
grad-cam>=1.4.0
captum>=0.6.0

# Config & Utils
pyyaml>=6.0
tqdm>=4.65.0
wandb>=0.15.0

# Dataset Download
requests>=2.31.0
gdown>=4.7.0

# Development
pytest>=7.3.0
black>=23.3.0
flake8>=6.0.0
jupyter>=1.0.0
ipykernel>=6.0.0
